pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tavily-cli"
version = "0.1.0"
description = "CLI and agent tools for the Tavily API — search, extract, crawl, map, and research from the command line."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Tavily", email = "dean.sacoransky@tavily.com" }
]
keywords = ["ai", "tavily", "web-search", "llm", "rag", "agents", "research", "crawl", "extract"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "tavily-python>=0.7.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://tavily.com"
Documentation = "https://docs.tavily.com"
Repository = "https://github.com/tavily-ai/skills"
Issues = "https://github.com/tavily-ai/skills/issues"

[project.scripts]
tvly = "tavily_cli.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["tavily_cli"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]
